Web format, huge policy and request classes update

+++ app/Http/Controllers/ProjectController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\Projet\CreateProjectRequest;
+use App\Http\Requests\Projet\UpdateProjectRequest;
 use App\Models\Project;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -22,17 +25,24 @@ class ProjectController extends Controller
         ]);
     }
 
+    /**
+     * @param Project $project
+     * @throws AuthorizationException
+     */
     public function show(Project $project)
     {
+        $this->authorize('view', $project);
+
         return view('pages.project-show', compact('project'));
     }
 
     /**
+     * @param CreateProjectRequest $request
      * @return RedirectResponse
      */
-    public function store(): RedirectResponse
+    public function store(CreateProjectRequest $request): RedirectResponse
     {
-        $validated = $this->validation(\request());
+        $validated = $request->validated();
         $validated['creator_id'] = auth()->id();
         Project::create($validated);
 
@@ -42,43 +52,42 @@ class ProjectController extends Controller
     /**
      * @param Project $project
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function destroy(Project $project): RedirectResponse
     {
+        $this->authorize('delete', $project);
+
         $project->delete();
 
         return redirect()->route('projects.index');
     }
 
+    /**
+     * @param Project $project
+     * @throws AuthorizationException
+     */
     public function edit(Project $project)
     {
+        $this->authorize('update', $project);
+
         return view('includes.project-edit', compact('project'));
     }
 
     /**
+     * @param UpdateProjectRequest $request
      * @param Project $project
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
-    public function update(Project $project): RedirectResponse
+    public function update(UpdateProjectRequest $request, Project $project): RedirectResponse
     {
-        $validated = $this->validation(\request());
+        $this->authorize('update', $project);
+
+        $validated = $request->validated();
         $validated['creator_id'] = auth()->id();
         $project->update($validated);
 
         return redirect()->route('projects.index');
-    }
-
-    /**
-     * @param Request $request
-     * @return array
-     */
-    public function validation(Request $request): array
-    {
-        return $request->validate([
-            'name' => 'required|max:50',
-            'description' => 'max:500',
-            'timeEst' => 'nullable|after_or_equal:today',
-            'effort' => 'nullable',
-        ]);
     }
 }
